Split attention_op out
